+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changes
 * Carthage bootstrap is executed also when only the Carthage.private file exists
 * Remove [Rome](https://github.com/tmspzz/Rome) support
 * Change that the altool use the apiKey and apiIssue. (username and password should still work)
+* Change that app extensions gets only the keychain-access-group of the main bundle set.
 
 # 0.20.1 (2. Dec, 2020)